morphedit as edit tool
