+++ packages/opencode/src/tool/edit-morph.txt
@@ -1,4 +1,4 @@
-Use this tool to make an edit to an existing file. 
+🚀 MORPH FAST APPLY - Use this tool to make an edit to an existing file.
 
 This will be read by a less intelligent model, which will quickly apply the edit. You should make it clear what the edit is, while also minimizing the unchanged code you write.
 When writing the edit, you should specify each edit in sequence, with the special comment // ... existing code ... to represent unchanged code in between edited lines.
