+++ packages/opencode/src/tool/registry.ts
@@ -1,7 +1,6 @@
 import z from "zod"
 import { BashTool } from "./bash"
 import { EditTool } from "./edit"
-import { MorphEditTool } from "./morphedit"
 import { GlobTool } from "./glob"
 import { GrepTool } from "./grep"
 import { ListTool } from "./ls"
@@ -19,7 +18,6 @@ export namespace ToolRegistry {
     InvalidTool,
     BashTool,
     EditTool,
-    MorphEditTool,
     WebFetchTool,
     GlobTool,
     GrepTool,
@@ -47,21 +45,21 @@ export namespace ToolRegistry {
     if (providerID === "openai") {
       return result.map((t) => ({
         ...t,
-        parameters: optionalToNullable(t.parameters),
+        parameters: optionalToNullable(t.parameters as z.ZodTypeAny),
       }))
     }
 
     if (providerID === "azure") {
       return result.map((t) => ({
         ...t,
-        parameters: optionalToNullable(t.parameters),
+        parameters: optionalToNullable(t.parameters as z.ZodTypeAny),
       }))
     }
 
     if (providerID === "google") {
       return result.map((t) => ({
         ...t,
-        parameters: sanitizeGeminiParameters(t.parameters),
+        parameters: sanitizeGeminiParameters(t.parameters as z.ZodTypeAny),
       }))
     }
 
@@ -91,16 +89,6 @@ export namespace ToolRegistry {
     if (modelID.includes("qwen")) {
       result["todowrite"] = false
       result["todoread"] = false
-    }
-
-    // If MORPH_API_KEY exists, only enable morphedit and disable other edit tools
-    if (process.env.MORPH_API_KEY) {
-      result["edit"] = false
-      result["multiedit"] = false
-      result["patch"] = false
-    } else {
-      // If no MORPH_API_KEY, disable morphedit tool
-      result["morphedit"] = false
     }
 
     return result
